requirements.txt
absl-py==2.3.1
asgiref==3.11.0
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deepface==0.0.96
Django==6.0
filelock==3.20.2
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
<<<<<<< HEAD
keras==3.13.0
=======
>>>>>>> bd1274c (Added Chat and rafactored code)
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.2
psycopg2==2.9.11
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
<<<<<<< HEAD
scikit-learn==1.6.1
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
=======
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
>>>>>>> bd1274c (Added Chat and rafactored code)
openpyxl==3.1.5
PyPDF2==3.0.1
python-docx==1.1.2
matplotlib==3.10.0
seaborn==0.13.2
tqdm==4.67.1
joblib==1.5.3
pyod==2.0.6
xgboost==2.1.3
lightgbm==4.6.0
<<<<<<< HEAD
nltk==3.9.1
spacy==3.8.3
plotly==5.24.1
=======
>>>>>>> bd1274c (Added Chat and rafactored code)
